requirements.txt
pandas~=2.2.3
requests~=2.32.3
openpyxl
python-dotenv~=1.1.0
openai~=1.74.0